Moves FixedMenu into 'src/components'
