+++ src/components/FixedMenu.js
@@ -6,8 +6,8 @@ import {
   Image
 } from 'semantic-ui-react';
 import { Link } from 'react-router-dom';
-import { BetaWarning } from './components/BetaWarning';
-import logo from './images/logo.svg';
+import { BetaWarning } from './BetaWarning';
+import logo from '../images/logo.svg';
 
 const FixedMenu = () => (
   <Menu fixed='top' size='large' color='teal' inverted >
